Rename ping router and remove redundant pytest.ini
