+++ laa_court_data_api_app/main.py
@@ -2,7 +2,7 @@ import logging
 
 import uvicorn
 from fastapi import FastAPI
-from routers import something
+from .routers import ping
 
 logging.config.fileConfig('logging.conf', disable_existing_loggers=False)
 
@@ -10,7 +10,7 @@ logger = logging.getLogger(__name__)
 
 app = FastAPI()
 
-app.include_router(something.router)
+app.include_router(ping.router)
 
 
 @app.get('/')
