+++ test/main_test.py
@@ -1,6 +1,6 @@
 from fastapi.testclient import TestClient
 from laa_court_data_api_app.main import app
-from laa_court_data_api_app.routers.something import router
+from laa_court_data_api_app.routers.ping import router
 from laa_court_data_api_app.config.app import get_app_settings
 import os
 
@@ -11,13 +11,16 @@ def test_root_returns_200():
     response = client.get('/')
     assert response.status_code == 200
 
+
 def test_root_returns_correct_body():
     response = client.get('/')
     assert response.json() == {'Status': 'Working'}
 
+
 def test_ping_returns_200():
     response = client.get('/ping')
     assert response.status_code == 200
+
 
 def test_ping_returns_correct_body():
     get_app_settings.cache_clear()
@@ -29,7 +32,7 @@ def test_ping_returns_correct_body():
 
     response = client.get('/ping')
     expected_result = {'app_branch': 'test_branch',
-               'build_date': '02022022',
-               'build_tag': 'test',
-               'commit_id': '123456'}
+                       'build_date': '02022022',
+                       'build_tag': 'test',
+                       'commit_id': '123456'}
     assert response.json() == expected_result
